refactor: avoid loops in PostBody
