+++ src/components/Posts/EncryptedPost.tsx
@@ -5,7 +5,7 @@ import { forwardRef, type HTMLProps } from 'react';
 
 import Lock from '@/assets/lock.svg';
 import { ClickableButton } from '@/components/ClickableButton.js';
-import { PostBody, type PostBodyProps } from '@/components/Posts/PostBody.js';
+import { PostBodyContent, type PostBodyContentProps } from '@/components/Posts/PostBodyContent.js';
 import { classNames } from '@/helpers/classNames.js';
 import { useDecryptPost } from '@/hooks/useDecryptPost.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
@@ -14,7 +14,7 @@ interface EncryptedPostProps extends HTMLProps<HTMLDivElement> {
     post: Post;
     isQuote?: boolean;
     disableDecrypt?: boolean;
-    postBodyProps?: Partial<PostBodyProps>;
+    postBodyProps?: Partial<PostBodyContentProps>;
 }
 
 export const EncryptedPost = forwardRef<HTMLDivElement, EncryptedPostProps>(function EncryptedPost(
@@ -24,27 +24,26 @@ export const EncryptedPost = forwardRef<HTMLDivElement, EncryptedPostProps>(func
     const [{ loading, value: decryptedPost }, decryptPost] = useDecryptPost(post);
 
     if (!loading && decryptedPost) {
-        return <PostBody {...postBodyProps} post={decryptedPost} disableDecrypt />;
+        return <PostBodyContent {...postBodyProps} post={decryptedPost} disableDecrypt />;
     }
 
     return (
         <div className={className} ref={ref}>
             <div
                 className={classNames(
-                    'flex items-center gap-1 rounded-lg border-primaryMain px-3 py-[6px] text-medium',
+                    'flex items-center justify-between rounded-lg border-primaryMain px-3 py-[6px] text-medium',
                     {
                         border: !isQuote,
                     },
                 )}
             >
-                <Lock width={16} height={16} />
-                {loading ? <Trans>Post is decrypting...</Trans> : <Trans>Post has been encrypted</Trans>}
+                <div className="flex items-center gap-1">
+                    <Lock width={16} height={16} />
+                    {loading ? <Trans>Post is decrypting...</Trans> : <Trans>Post has been encrypted</Trans>}
+                </div>
                 {!loading && !disableDecrypt && post.canDecrypt ? (
-                    <ClickableButton
-                        onClick={decryptPost}
-                        className="p-1 text-sm font-semibold text-highlight hover:bg-highlight/25"
-                    >
-                        <Trans>Try decrypt</Trans>
+                    <ClickableButton onClick={decryptPost} className="text-base text-highlight">
+                        <Trans>Decrypt</Trans>
                     </ClickableButton>
                 ) : null}
             </div>
