+++ src/hooks/useDecryptPost.ts
@@ -4,7 +4,7 @@ import { useAccount } from 'wagmi';
 
 import { Source } from '@/constants/enum.js';
 import { ENABLED_DECRYPT_SOURCES } from '@/constants/index.js';
-import { enqueueMessageFromError } from '@/helpers/enqueueMessage.js';
+import { enqueueMessageFromError, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
 import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
 import { resolveSourceName } from '@/helpers/resolveSourceName.js';
 import { ConnectModalRef } from '@/modals/controls.js';
@@ -23,7 +23,12 @@ export function useDecryptPost(post: Post) {
             }
 
             const provider = resolveSocialMediaProvider(post.source);
-            return await provider.decryptPost(post);
+            const result = await provider.decryptPost(post);
+            if (result) {
+                enqueueSuccessMessage(t`Post decrypted successfully!`);
+            }
+
+            return result;
         } catch (error) {
             enqueueMessageFromError(error, t`Failed to decrypt post on ${resolveSourceName(post.source)}.`);
             return null;
